+++ changelog.md
@@ -7,6 +7,7 @@
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/312 (!include doesn't work with files encoded as UTF-8 BOM).
 - structurizr-dsl: Adds a way to explicitly specify the order of relationships in dynamic views.
 - structurizr-dsl: Adds support for element technology expressions (e.g. "element.technology==Java").
+- structurizr-dsl: Adds an `!elements` keyword that can be used to find a set of elements via an expression.
 
 ## 2.2.0 (2nd July 2024)